Update process.go (#530)
We use a product called metabase and previously we could deploy this nice little convox.yml:

```yml
resources:
    database:
      type: postgres

services:
  metabase:
    image: metabase/metabase:v0.44.6
    environment:
      - JAVA_TOOL_OPTIONS=-Xmx1g
    port: 3000
    resources:
      - database:MB_DB_CONNECTION_URI
```

But after #481 it will not work anymore because it changed that the exposed env var from a resource from `URL` to whatever is at the end of the env var we want to map it to. 

I don't think that was intentional as the "contract" for a resource seemed to be that it exposes a `URL` env var.